backend/requirements.txt
alembic
fastapi
psycopg2-binary
python-dotenv
requests
sqlalchemy
uvicorn